+++ Backend/Services/LuaServiceLib/LuaContext.cs
@@ -26,6 +26,7 @@ namespace Slipstream.Backend.Services.LuaServiceLib
                 UIMethodCollection.Register(eventBus, eventFactory, Lua);
                 InternalMethodCollection.Register(eventBus, eventFactory, Lua);
                 HttpMethodCollection.Register(eventBus, eventFactory, Lua);
+                IRacingMethodCollection.Register(eventBus, eventFactory, Lua);
 
                 // Fix paths, so we can require() files relative to where the script is located
                 var ScriptPath = Path.GetDirectoryName(filePath).Replace("\\", "\\\\");