+++ Shared/EventHandler.cs
@@ -128,6 +128,9 @@ namespace Slipstream.Shared
         public delegate void OnIRacingPitstopReportHandler(EventHandler source, EventHandlerArgs<Shared.Events.IRacing.IRacingPitstopReport> e);
         public event OnIRacingPitstopReportHandler? OnIRacingPitstopReport;
 
+        public delegate void OnIRacingCommandSendCarInfoHandler(EventHandler source, EventHandlerArgs<Shared.Events.IRacing.IRacingCommandSendCarInfo> e);
+        public event OnIRacingCommandSendCarInfoHandler? OnIRacingCommandSendCarInfo;
+
         public delegate void OnIracingDriverIncidentHandler(EventHandler source, EventHandlerArgs<Shared.Events.IRacing.IRacingDriverIncident> e);
         public event OnIracingDriverIncidentHandler? OnIracingDriverIncident;
         #endregion
@@ -380,6 +383,13 @@ namespace Slipstream.Shared
                     else
                         OnIRacingPitstopReport.Invoke(this, new EventHandlerArgs<Shared.Events.IRacing.IRacingPitstopReport>(tev));
                     break;
+
+                case Shared.Events.IRacing.IRacingCommandSendCarInfo tev:
+                    if (OnIRacingCommandSendCarInfo == null)
+                        OnDefault?.Invoke(this, new EventHandlerArgs<IEvent>(tev));
+                    else
+                        OnIRacingCommandSendCarInfo.Invoke(this, new EventHandlerArgs<Shared.Events.IRacing.IRacingCommandSendCarInfo>(tev));
+                    break; 
 
                 case Shared.Events.IRacing.IRacingDriverIncident tev:
                     if (OnIracingDriverIncident == null)